Add lib.rs function mapping `stream` and `parse`
Alas, I had to ruin some of the beauty of the parse module for this.
Borrowing from the original line became too difficult, so we need
local ownership in the parsed objects.

The performance penalty is not _that_ big though, so I'd rather
live with it than try to come up with something clever.