Add `Filters` object and post parsing filters
If we're going to store the output directly to parquet or other
formats, it will be very useful to be able to filter both before
parsing (regex) and after (object).

I just added `mobile` for now to test it, but will add more later.
We also need better python bindings for this.